requirements.txt
watchdog
google-generativeai